refactor: 모임 운동 관련 dto를 별도 패키지로 분리
