+++ src/main/java/umc/cockple/demo/domain/exercise/converter/query/PartyExerciseCalendarQueryMapper.java
@@ -2,7 +2,7 @@ package umc.cockple.demo.domain.exercise.converter.query;
 
 import org.springframework.stereotype.Component;
 import umc.cockple.demo.domain.exercise.domain.Exercise;
-import umc.cockple.demo.domain.exercise.dto.PartyExerciseCalendarDTO;
+import umc.cockple.demo.domain.exercise.dto.party.PartyExerciseCalendarDTO;
 import umc.cockple.demo.domain.party.domain.Party;
 import umc.cockple.demo.global.enums.Gender;
